s3:librpc/rpc: split out dcerpc_internal.h for struct pipe_auth_data
This should be anonymous for callers.

(cherry picked from commit b874619b5ed5d29a939f13d8064a47c8aca45b8c)
