+++ source3/rpc_client/cli_pipe.c
@@ -31,6 +31,7 @@
 #include "librpc/gen_ndr/ndr_netlogon_c.h"
 #include "librpc/gen_ndr/auth.h"
 #include "librpc/rpc/dcerpc.h"
+#include "librpc/rpc/dcerpc_internal.h"
 #include "librpc/rpc/dcerpc_util.h"
 #include "rpc_dce.h"
 #include "cli_pipe.h"
